Tweak paging label. Removed call to detachNewThread by placing statement after async call.

+++ wikibrowser/NSString+UrlEncodeForWiki.m
@@ -16,7 +16,7 @@
 - (NSString *)urlEncodeForWiki {
     NSMutableString *output = [NSMutableString string];
     const unsigned char *source = (const unsigned char *)[self UTF8String];
-    int sourceLen = strlen((const char *)source);
+    unsigned long sourceLen = strlen((const char *)source);
     for (int i = 0; i < sourceLen; ++i) {
         const unsigned char thisChar = source[i];
         if (thisChar == ' '){
